do not call kvm upgrade
kvm upgrade is not needed because omnisharp explicitly needs 1.0.0-beta2. kvm upgrade sets the latest version of KVM as 'default' forcing VS to pick it up. so if you run the build.cmd you will not be able to compile in VS until you manually reset the 'default' back to 1.0.0-beta2