Added initial run of typescript generation.
Fixed build.sh (in theory), fixed project.json

:shaved_ice: \ != /

Updates to create better typings for consumption in the client

Change around how methods are generated... this makes it a little easier to consume from the typescript side.

Added observers to the interface as well

Added promises to the ts output

Modified obserables to include the request.

ICollection was being missed for collection of projects on quick fix

Typescript Naming things is hard