Fixes OmniSharp/omnisharp-roslyn#272
On my system I found that 4 seconds was not logn enough for DHT to start.
Time for a new computer perhaps...

Anyway I have bumped the timeout to 10 seconds (FYI: my system on average took
6 seconds to start DHT).

I have also added in a check to see if the timeout has indeed been reached.