Conditionally apply changes for rename server side.
Added a new flag to the request. Ideally, we would have used
WantsTextChanges to both return text changes and also to not apply
them at immediately at the server.

The assumption here is that new text changes will be returned to the
server as they are applied to the editor.