Fix span used to determine whether signature help should be shown
OmniSharp used the full span of an argument list to determine whether signature help should be shown. However, that includes both leading and trailing trivia, which means that it would return signatures for positions outside of the argument list. In addition, I switched it to use `Contains(...)` rather than `IntersectsWith(...)` since checking the position against the argument list span should not include the outer positions (e.g. on the left or right of the parentheses).