Fix project.json search in DotNetProjectSystem
This change fixes several issues with locating project.json projects:

1. The top-level folders defined in global.json were not searched. So, if
a user tried to add `"projects": ["MyProject1"]` where `MyProject1`
contained a project.json file, the project wouldn't be found. Instead,
only child folders of the top-level folders would be search. So,
`src/MyProject1` would work. Now, we'll search the top-level folders and
their immediate children, which should match how `dotnet restore` behaves
a bit more closely. A test has been added for this scenario.

2. We had infrastructure for falling back to a recursive directory search
when a project.json or global.json doesn't exist in the current directory.
However, we weren't using it! There was an issue with the recursive search
(it wouldn't search a directory's sub-directories if the directory
contained a project.json file), but that is now fixed.

3. The test for recursively searching wasn't correct. This is now fixed.