Use MSBuild solution parsing API
This change removes the old custom solution parser that had been copied
from Roslyn and uses the MSBuild solution parsing API instead. In
addition, because the MSBuild API does not provide access to the actual
'project type' GUID specified in the solution file, detection of Unity
projects is handled differently. Rather than hashing the project name
and seeing if it matches the 'project type' GUID, we just look look for
the presence of 'UnityEngine.dll' in a project's references.