Fix build script to publish for all RIDs
Awhile back, I had added code to the build script to avoid publishing
multiple times for the same RID. However, that failed to take into
account the "default" RID, which is intended to represent the current
machine. When we encounter "default", we substitute the RID for the
current machine, which means that we'd never actually publish that same
RID. This resulted in builds being missing if the RID matched the
current machines. In particular, this resulted in ubuntu 14 builds being
missing from Travis because that's the "default".