Increase the timeout specified when running published bits during build
Our build script finishes up by running the published OmniSharp bits with a timeout. This timeout is no longer long enough and the bits continue trying to run after the build is finished. I *believe* the reason for this is because the MSBuildProjectSystem now loads referenced projects outside of the directory OmniSharp is started with. More projects means more time processing, which causes the bits run longer. I'm increasing the timeout from 10 seconds to 30 seconds.