Stop setting MSBuildSDKsPath environment variable per project
With the SDK Resolvers included in OmniSharp, we no longer need to set the MSBuildSDKsPath
environment variable for each project. Instead, the SDK Resolvers should do that work.

For tests, we now ensure that MSBuild is always initialized in standalone mode and we
set the MSBuildSDKsPath environement variable specifically to ensure that the .NET Core SDK(s)
downloaded by the OmniSharp build are used in tests.