Re-organize MSBuild folder
This change re-organizes the layout of the MSBuild folder to more closely
match the layout of Visual Studio:

* "msbuild" - This will be the MSBuildExtensionsPath
* "msbuild\15.0" - common props go here
* "msbuild\15.0\bin" - The MSBuild tools path, where much of the MSBuild actually goes.

In addition, the MSBuild runtime and library packages have been updated to the latest,
and the package references for the libraries are marked as `PrivateAssets="all"` in the
OmniSharp.MSBuild project. This last change means that OmniSharp no longer runs, since the
MSBuild libraries aren't copied to the OmniSharp base application path any longer. Later
commits will fix this.