Fix issue with C# projects referencing F# projects
This fixes a long-standing problem where C# projects would attempt to load project references
to non-C# projects. This is related to the problem on not adding file references for C# project references
to avoid duplicate references. Now, we're a bit more selective when we decide to throw out a project or file
reference: We check the file path and see if it ends with ".csproj".