Fix loading of Xamarin.Android solution
Xamarin.Android solution (Xamarin.Android.sln in
https://github.com/xamarin/xamarin-android/) fails to load with the current
version of omnisharp-roslyn because, somehow, one of the project reference file
paths is null or empty. I don't know which project that is (there are around
1500 of them in the XA repo) and, unfortunately, I had no time to investigate
that further. However, the change in this diff makes it work and I think it's a
good safeguard to have this check in place.