Update OmniSharp to latest MSBuild
This change updates OmniSharp with the latest MSBuild binaries. Recent performance changes for MSBuild now
require NuGet.Build.Tasks, which requires a set of NuGet binaries be included as well. Because the versions of these NuGet binaries might be different (actually, they *are* different) than the binaries that OmniSharp uses, there's a bit of duplication.