Fix subtle issue that breaks handling of discovered referenced projects
Recently, we added some additional logging that dumps the ProjectFileInfo using
Json.net when a project is added for debugging purposes. Unfortunately, this throws
when the ProjectFileInfo was created for a referenced project that is outside of
the current solution or OmniSharp target directory. The exception thrown is an
InvalidOperationException during Json.net serialization because several of the
properties on the ProjectFileInfo are "unintialized" `ImmutableArray<T>` values.
This happens specifically for project references outside of the solution/targetdir
because OmniSharp just creates a barebones ProjectFileInfo and queues an update to
fill it with real values later.

I'm fixing the issue be ensuring that ProjectFileInfo `ImmutableArray<T>` properties
are always initialized to `ImmutableArray<T>.Empty`. I've also added a test to cover
the particular scenario of referenced projects outside of the solution/targetdir. We
didn't have a test already, which is why this wasn't caught sooner.