Set MSBuild property to allow the XAML markup compiler task to run
We need to set the AlwaysCompileMarkupFilesInSeparateDomain property to false.
Otherwise, the XAML markup compiler task will try to create an AppDomain to run
the markup compiler in. In normal MSBuild scenarios, this is fine. However, in
OmniSharp, our AppDomain.AssemblyResolve handler that's used to locate MSBuild
will not be installed in the marker compiler task's AppDomain, causing it to fail.