Ensure that we don't subscribe the same file-watcher callback multiple times
While refactoring some code, I noticed that we were queuing the same callback with the
`IFileSystemWatcher` many, many times in order to watch source file directories. That means
that these callbacks would be called multiple times for any notification. This change expands
the `ManualFileSystemWatcher` to only add the same callback for a particular extension, file
or directory once. Note that the callback is only checked for reference equality, so it's
necessary to stash your callback in a field if you don't want it added multiple times.