Don't set MSBuildSDKsPath environment variable by default
This commit removes an old hack in the MSBuild project loader that blocks using
custom MSBuild SDKs.

Details:

The MSBuild project loader has a bit of a hack to try and find the correct
.NET Core SDK for a given project. Essentially, it tries to launch "dotnet --info"
in the project directory and sets the MSBuildSDKsPath using the path information
that's returned. This results in several problems:

1. "dotnet --info" gets run for *every* project that gets loaded, even if they
   aren't SDK-style projects.

2. This hack overrides custom MSBuild SDK resolvers. This worked OK while there
   really weren't custom SDK resolvers out there. However, that is no longer the
   case with MSBuild's new NuGet-based SDK resolver. This SDK resolver is used
   to acquire and use custom MSBuild SDKs. So, this hack blocks that feature.

OmniSharp does provide a default SDK resolver. However, this hack was left in place
because that SDK resolver has problems with discovering much older .NET Core SDKs,
i.e. <= 1.0.3. It's unlikely that many users are still on such old versions, but for
those who are, a new option has been added to restore the hack.