Rationalize IDotNetCliService
This change moves the DotNetCliService from a singleton class instantiated via MEF
to a service that is added to OmniSharp's IServiceProvider. The is a new IDotNetCliService
interface in OmniSharp.Abstractions and the DotNetCliService implementation has been moved
to OmniSharp.Host. This greatly cleans up how DotNetCliService is instantiated by tests and
allows the .NET CLI path to be precomputed for testing.

In addition, IEventEmitter is now elevated to a required service in the IServiceProvider.