Fix bug when multi-TFM project reference is outside OmniSharp's target path
When OmniSharp loads a project reference outside the OmniSharp path, it
does not attempt to do a full design-time build of the project. Instead,
it just uses the MSBuild project evaluation. However, an important code path
was being skipped where a target framework is selected in the case that the
project has multiple target frameworks. Later, this causes an exception to
be thrown because several properties did not evaluate properly. This change
ensures that a target framework is selected in this case.