Add sink API to notify when MSBuild projects are loaded
This change introduces a new API that can be MEF exported by OmniSharp plug-ins in order to be
notified when the MSBuild project loads or reloads a project. This API provides access to the
Microsoft.Build.Evaluation.ProjectInstance that was built so that consumers (e.g. Razor) can
build additional targets if necessary.

Note that any plug-in using this API will need to add a package reference to Microsoft.Build like so:

```XML
<PackageReference Include="Microsoft.Build" ExcludeAssets="Runtime" PrivateAssets="all" />
```