Fix OmniSharp log filtering
When the logging packages were updated from 1.1 to 2.1, we took a breaking change
(https://github.com/aspnet/Announcements/issues/238) that caused the --loglevel switch to no longer work properly. This
change uses the new APIs for configuring logging and centralizes the filter. Note that I didn't change the LSP implementation
because it looks to me like LSP is using loggers differently.