Add RunTestsInContextComment
This command tries to replicate the ctrl+R, T experience in full VS. I'm planning on sending a followup PR later in the week to add DebugInContext, and then a PR to the vscode extension to enable the extension. While Omnisharp does have an existing TestCommandService that does some of this, it doesn't quite replicate the experience: it will only run the test method that it's invoked in, not potentially the entire test class or whatever other context it's in. Additionally, it's not unified with the test running framework the test of the V2 test commands use, so I opted to introduce an entirely new command, RunTestsInContext, which will walk up the syntax tree for a given location to find the nearest test method or class, if one exists, and either run the method itself, or run the entire class.