Start moving omnisharp to directly using Roslyn's completion service, with a separate resolve step
Today, Omnisharp uses Roslyn's completion service to gather completions, but doesn't then use the other completion helpers to fill in the rest of the detail of the completion. Additionally, Omnisharp does a bunch of work when creating and populating completions, and this can cause slow and inconsistent behavior when invoking completions in the middle of a pre-typed word. Some current issues:

* vscode fuzzy-matching changes based on whether you invoke completion after typing a few characters.
* Completions with large numbers of elements can take a bit of time to load, as Omnisharp does filtering and documentation building for all of them.
* Completion is not well setup to play nicely with LSP, which uses a much different interface.
* Completion is not capable of providing unimported types, particularly because providing the edits would be extremely expensive if not done only on request.

To solve these issues, this adds a new completion service that mirrors the structure of LSP completion and completion/resolve:

* On request, only the completion item and initial edit are provided.
* A new response service is added, which looks up documentation when a particular item is selected in the completion window.
    * This response service will additionally be able to provide extra edits, such as inserting the rest of an override statement or adding imports for unimported types, but these are not yet implemented.