Implement correct override, partial, xml doc tag, and internalsvisibleto completion.
LSP does not support a concept of asynchronous edits that touch the current cursor location: anything that does needs to advertise those edits up front. That makes it impossible to do override or partial completion lazily: completing the method body _requires_ touching the insertion text, as depending on editorconfig settings braces might need to be inserted at the end of the line. Further, moving the cursor requires using a snippet, which cannot be accomplished in an additionalTextEdit. So, for those 4 providers that need to compute additional work at the current cursor location, we do so. We take any work that comes before the cursor location (such as turning override into public override returntype) and turn that into an additionalTextEdit, and then snippitize the rest of the completion and use it as insertionText.