Refactor the MarkdownHelper.
This allows it to be fully shared by the QuickInfoProvider and the new CompletionService, with much less special casing and encapsulation violations.

Additionally, the responses from the CompletionService now no longer specify the same lower-cased names as the the LSP completion endpoint. This brings the service into consistency with the rest of omnisharp, and it won't be very hard for the LSP provider to wrap it.