Better handle completion when the display text is not in the final result
Override and partial method completion can have scenarios where the DisplayText of the completion item does not appear in the final text, such as when types need to be imported or when nullability differs between original declaration and generation declaration sites. We now handle these by parsing the change if we fail to find the insertion text, and using the last method in the change to find the correct location in the new text. Fixes https://github.com/OmniSharp/omnisharp-roslyn/issues/1907.