Correctly handle <ProjectReferences> that don't produce references
<ProjectReference> elements can have metadata that specify what type
of item to create, or that they shouldn't create items at all.
The existing code that was processing references wasn't checking for
this metadata, so you could end up with project references between
projects that shouldn't be there.

The approach to take here is to never look at the <ProjectReference>
items directly, rather only look at the ReferencePath items that are
produced, and if we see metadata that indicates it came from a project,
convert it to a project reference at that point.