Adds SyntaxTree Visualizer Endpoints
This adds a new set of endpoints designed for implementing a syntax tree visualizer. There is an accompanying set of commits using these endpoints in vscode to display syntax information for the currently-active file. This is a stateful API that requires calls in the correct order. A session can be started by querying one of 2 endpoints:

1. SyntaxTreeRequest with a null Parent.
2. SyntaxNodeAtRangeRequest with a file path that does not match the current path and text version.

The rest of the tree can be filled in by repeated requests to SyntaxTreeRequest with the parent filled in, in order to get the child nodes at that level.

It is also useful to build the tree in reverse order, starting from a nested node and working up. This enabled a combination of SyntaxNodeAtRangeRequest followed by repeated SyntaxNodeParentRequests to build up the spine of the tree, which allows clients to figure out what node to reveal in the tree when a document range is selected.

Finally, SyntaxNodeInfoRequest returns detailed information about a particular tree item. It returns the node type (and the symbol kind of that type), the SyntaxKind, the semantic classification of the node, all public properties of the node, and if the node is a SyntaxNode, the results of calling GetTypeInfo, GetConversion, GetSymbolInfo, and GetDeclaredSymbol on that node, enabling a detail view of the node.