Remove private reflection from OmniSharp
This updates OmniSharp to consume the new ExternalAccess assemblies that were added to Roslyn, providing a more structured form of internals access that will give us compile-time diagnostics and avoid the overhead of reflection-based calls. As a part of this, I'm updating our Roslyn dependency to 3.11, and I'm removing the deprecated AutoComplete handler in favor of the modern completion handlers.