Add V2 version of GotoDefinitionService
The V1 GotoDefinitionService can only return one definition for a symbol. This is particularly annoying for partial types, which have multiple locations, and the location the service returns is pretty much never the one I want. The V2 service is very similar, but it uses V2 location models and returns a list of definitions, not a single one.