Add a configuration mechanism to allow users to point Omnisharp to a custom .NET CLI, with a fallback to the DOTNET_ROOT environment variable before falling back to the PATH.
I've updated the tests to use this functionality for locating the .NET CLI instead of using the old test-only code-path.

As this is my first contribution to Omnisharp, I'm not quite sure how this exposes the new setting through to omnisharp.json, so if anyone has any pointers on that, they would be well appreciated.