Only get first document's highlights
For files that are included in multiple projects (such as via `Compile Include` directives), we currently attempt to retrieve semantic highlights for every document. However, this means that highlight spans will have overlapping results from every project, and the service will crash. Further, we have no way of actually selecting what project to display the file as, so retrieving any but the first document doesn't serve the user. Therefore, we remove the loop, and just look through the first document found.